Address PR #268 review (Copilot + Sonar)
Copilot inlines:
- policy-editor.spec.ts:83 — handleCreateRule returns 201 Created, not
  200; flip the assertion so the test pins the real wire contract
  (server/rules/internal/operator/appcontrol_handler.go:233).
- rule-detail.spec.ts:13 — stale "/api/rules/docs" path in the file
  header comment; the code calls /api/rules. Update the comment.

Sonar issues:
- typescript:S7721 process-tree-detail.spec.ts — move the
  gotoTreeWithProcesses helper out of the describe(). Replaced by a
  shared module fixture (see below).
- typescript:S7780 rule-detail.spec.ts:61 — switch the severity-row
  regex to String.raw so \s isn't backslash-escaped.

Sonar gate (new_duplicated_lines_density = 3.4% vs 3% threshold):
- Extract test/e2e/fixtures/process-tree.ts with setupProcessTreeDeep,
  the shared "enrol host + post process-tree-deep + wait for processor
  + navigate to /ui/hosts/<id> + wait for tree" helper. Both
  host-list-and-process-tree.spec.ts and process-tree-detail.spec.ts
  now delegate to it instead of inlining the same ~25-line block.
- Refactor tests/auth/oidc-login.spec.ts to call a module-level
  signInAsAnalystViaDex helper instead of inlining the dex ceremony in
  each test() block. Local helper (not a copy of _setup.signInViaDex)
  so the auth/* suite stays independent of the qa/* setup graph.
